rise_app_backend/requirements.txt
langgraph>=0.2.0
openai
django>=4.0
python-dotenv
fastmcp
aiohttp